setup.cfg: Enforce 100% branch coverage
Creates a list of coverage exceptions in .nocover.yaml
which are used in setup.cfg by coverage to ignore them
while ensuring any other parts of the code are fully
tested.

Also rename one function so its signature is on one line
as required for coverage to ignore the entire function.

Closes https://github.com/coala/coala-quickstart/issues/282